refactor(llms): colocate reasoning controls
